requirements.txt
bleak>=0.21.1
PyYAML>=6.0

